+++ bestmobabot/bot.py
@@ -558,10 +558,8 @@ class Bot(contextlib.AbstractContextManager, BotHelperMixin):
         """
         logger.info('🎁 Farming offers…')
         for offer in self.api.get_all_offers():
-            if not offer.is_free_reward_obtained:
-                try:
-                    log_reward(self.api.farm_offer_reward(offer.id))
-                except NotFoundError:
-                    logger.warning(f'🤔 Offer #{offer.id} is not found.')
+            logger.debug(f'🎁 #{offer.id}: {offer.offer_type}.')
+            if offer.offer_type == 'sideBarIcon_tripleSkinBundle' and not offer.is_free_reward_obtained:
+                log_reward(self.api.farm_offer_reward(offer.id))
             else:
                 logger.info(f'🎁 #{offer.id}: free reward is already obtained.')